limitado acceso de cliente en los controladores

+++ app/Http/Controllers/ClienteController.php
@@ -28,9 +28,14 @@ class ClienteController extends Controller
      */
     public function index()
     {
-        $clientes= Cliente::get();
+        if(auth()->user()->tipo === 'empleado'){
+            $clientes= Cliente::get();
 
-        return view('clientes.index', compact( 'clientes'));
+            return view('clientes.index', compact( 'clientes'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -40,7 +45,12 @@ class ClienteController extends Controller
      */
     public function create()
     {
-        return view('clientes.create');
+        if(auth()->user()->tipo === 'empleado'){
+            return view('clientes.create');
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -51,54 +61,57 @@ class ClienteController extends Controller
      */
     public function store(ClienteRequest $request)
     {
-        if($request->has('nombre')){
-            $user=new User();
-            $user->nombre=$request->get('nombre');
-            $user->apellido=$request->get('apellido');
-            $user->domicilio=$request->get('domicilio');
-            $user->dni=$request->get('DNI');
-            $user->email=$request->get('email');
-            $user->tel=$request->get('tel');
-            $user->fecha_nac=$request->get('fecha_nacimiento');
-            $user->datos=$request->get('datos');
-            $user->tipo="cliente";
-            if($request->hasFile('foto')){
-                $file = $request->file('foto');
-                $path = 'img/fotos/';
-                $filename= time().'-'.$file->getClientOriginalName();
-                $uploadSuccess = $request->file('foto')->move($path , $filename);
-                $user->foto = $path . $filename;
+        if(auth()->user()->tipo === 'empleado'){
+            if($request->has('nombre')){
+                $user=new User();
+                $user->nombre=$request->get('nombre');
+                $user->apellido=$request->get('apellido');
+                $user->domicilio=$request->get('domicilio');
+                $user->dni=$request->get('DNI');
+                $user->email=$request->get('email');
+                $user->tel=$request->get('tel');
+                $user->fecha_nac=$request->get('fecha_nacimiento');
+                $user->datos=$request->get('datos');
+                $user->tipo="cliente";
+                if($request->hasFile('foto')){
+                    $file = $request->file('foto');
+                    $path = 'img/fotos/';
+                    $filename= time().'-'.$file->getClientOriginalName();
+                    $uploadSuccess = $request->file('foto')->move($path , $filename);
+                    $user->foto = $path . $filename;
+                }
+                $pass = Str::random(8);
+                $user->password=bcrypt($pass);
+
+                $user->save();
+                $cl=new Cliente();
+                $cl->user_id=$user->id;
+                $cl->idioma=$request->idioma;
+                if($request->has('SIP')){
+                    $cl->sip=Crypt::encryptString($request->SIP);
+                }
+                $cl->nacionalidad=$request->nacionalidad;
+                if($request->has('enfermedades')){
+                    $cl->enfermedades=Crypt::encryptString($request->enfermedades);
+                }
+                $cl->familiares=$request->family;
+                $cl->coordenadax=$request->coordenadax;
+                $cl->coordenaday=$request->coordenaday;
+                $cl->user()->associate($user);
+                $cl->save();
+
+                $details=[
+                    'password' => $pass
+                ];
+                if($request->has('email') && $user->email != ""){
+                    Mail::to($user->email)->send(new PassMail($details));
+                }
+
+                return back()->with('info','Se ha creado el registro.');
             }
-            $pass = Str::random(8);
-            $user->password=bcrypt($pass);
-
-            $user->save();
-            $cl=new Cliente();
-            $cl->user_id=$user->id;
-            $cl->idioma=$request->idioma;
-            if($request->has('SIP')){
-                $cl->sip=Crypt::encryptString($request->SIP);
-            }
-            $cl->nacionalidad=$request->nacionalidad;
-            if($request->has('enfermedades')){
-                $cl->enfermedades=Crypt::encryptString($request->enfermedades);
-            }
-            $cl->familiares=$request->family;
-            $cl->coordenadax=$request->coordenadax;
-            $cl->coordenaday=$request->coordenaday;
-            $cl->user()->associate($user);
-            $cl->save();
-
-            $details=[
-                'password' => $pass
-            ];
-            if($request->has('email') && $user->email != ""){
-                Mail::to($user->email)->send(new PassMail($details));
-            }
-
-            return back()->with('info','Se ha creado el registro.');
-
-
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
         }
     }
 
@@ -110,9 +123,14 @@ class ClienteController extends Controller
      */
     public function show($id)
     {
-        $cliente= Cliente::findOrFail($id);
+        if(auth()->user()->tipo === 'empleado'){
+            $cliente= Cliente::findOrFail($id);
 
-        return view('clientes.show', compact('cliente'));
+            return view('clientes.show', compact('cliente'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -123,8 +141,13 @@ class ClienteController extends Controller
      */
     public function edit($id)
     {
-        $cliente= Cliente::findOrFail($id);
-        return view('clientes.edit', compact('cliente'));
+        if(auth()->user()->tipo === 'empleado'){
+            $cliente= Cliente::findOrFail($id);
+            return view('clientes.edit', compact('cliente'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -136,40 +159,44 @@ class ClienteController extends Controller
      */
     public function update(ClienteRequest $request, $id)
     {
-        $cl= Cliente::findOrFail($id);
-        $cl->user->nombre=$request->get('nombre');
-        $cl->user->apellido=$request->get('apellido');
-        $cl->user->domicilio=$request->get('domicilio');
-        $cl->user->dni=$request->get('DNI');
-        $cl->user->email=$request->get('email');
-        $cl->user->tel=$request->get('tel');
-        $cl->user->fecha_nac=$request->get('fecha_nacimiento');
-        $cl->user->datos=$request->get('datos');
-        $cl->user->fecha_baja=$request->get('fecha_baja');
-        if($request->hasFile('foto')){
-            $file = $request->file('foto');
-            $path = 'img/fotos/';
-            $filename= time().'_'.$file->getClientOriginalName();
-            $uploadSuccess = $request->file('foto')->move($path , $filename);
-            $cl->user->foto = $path . $filename;
-        }
-        $cl->user->save();
-        $cl->idioma=$request->idioma;
-        if($request->has('SIP')){
-            $cl->sip=Crypt::encryptString($request->SIP);
-        }
-        $cl->nacionalidad=$request->nacionalidad;
-        if($request->has('enfermedades')){
-            $cl->enfermedades=Crypt::encryptString($request->enfermedades);
-        }
-        $cl->familiares=$request->family;
-        $cl->coordenadax=$request->coordenadax;
-        $cl->coordenaday=$request->coordenaday;
-        $cl->save();
-        Session::flash('info', "Se han actualizado los datos.");
+        if(auth()->user()->tipo === 'empleado'){
+            $cl= Cliente::findOrFail($id);
+            $cl->user->nombre=$request->get('nombre');
+            $cl->user->apellido=$request->get('apellido');
+            $cl->user->domicilio=$request->get('domicilio');
+            $cl->user->dni=$request->get('DNI');
+            $cl->user->email=$request->get('email');
+            $cl->user->tel=$request->get('tel');
+            $cl->user->fecha_nac=$request->get('fecha_nacimiento');
+            $cl->user->datos=$request->get('datos');
+            $cl->user->fecha_baja=$request->get('fecha_baja');
+            if($request->hasFile('foto')){
+                $file = $request->file('foto');
+                $path = 'img/fotos/';
+                $filename= time().'_'.$file->getClientOriginalName();
+                $uploadSuccess = $request->file('foto')->move($path , $filename);
+                $cl->user->foto = $path . $filename;
+            }
+            $cl->user->save();
+            $cl->idioma=$request->idioma;
+            if($request->has('SIP')){
+                $cl->sip=Crypt::encryptString($request->SIP);
+            }
+            $cl->nacionalidad=$request->nacionalidad;
+            if($request->has('enfermedades')){
+                $cl->enfermedades=Crypt::encryptString($request->enfermedades);
+            }
+            $cl->familiares=$request->family;
+            $cl->coordenadax=$request->coordenadax;
+            $cl->coordenaday=$request->coordenaday;
+            $cl->save();
+            Session::flash('info', "Se han actualizado los datos.");
 
-        return redirect()->route('clientes.show', $id);
-
+            return redirect()->route('clientes.show', $id);
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -180,23 +207,28 @@ class ClienteController extends Controller
      */
     public function destroy($id)
     {
-        $cl= Cliente::findOrFail($id);
-        try
-        {
-            if(Servicio::where('cliente_id', '=', $id)->first() != null || Tratamiento::where('cliente_id', '=', $id)->first() != null
-                || Incidencia::where('cliente_id', '=', $id)->first() != null){
-                return back()->with('error','Se han creado registros con este cliente, no puede ser eliminado.');
-            }else{
-                $cl->user->delete();
-                $cl->delete();
-                Session::flash('info', "Se ha eliminado el registro.");
-                return redirect()->route('clientes.index');
+        if(auth()->user()->tipo === 'empleado'){
+            $cl= Cliente::findOrFail($id);
+            try
+            {
+                if(Servicio::where('cliente_id', '=', $id)->first() != null || Tratamiento::where('cliente_id', '=', $id)->first() != null
+                    || Incidencia::where('cliente_id', '=', $id)->first() != null){
+                    return back()->with('error','Se han creado registros con este cliente, no puede ser eliminado.');
+                }else{
+                    $cl->user->delete();
+                    $cl->delete();
+                    Session::flash('info', "Se ha eliminado el registro.");
+                    return redirect()->route('clientes.index');
+
+                }
+
+            }catch(Exception $e){
+                return $e->getMessage();
 
             }
-
-        }catch(Exception $e){
-            return $e->getMessage();
-
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
         }
     }
 }
