+++ app/Http/Controllers/EmpleadoController.php
@@ -28,10 +28,15 @@ class EmpleadoController extends Controller
      */
     public function index()
     {
+        if(auth()->user()->tipo === 'empleado'){
 
-        $empleados= Empleado::get();
+            $empleados= Empleado::get();
 
-        return view('empleados.index', compact( 'empleados'));
+            return view('empleados.index', compact( 'empleados'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -41,8 +46,12 @@ class EmpleadoController extends Controller
      */
     public function create()
     {
-
-        return view('empleados.create');
+        if(auth()->user()->tipo === 'empleado'){
+            return view('empleados.create');
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -53,46 +62,47 @@ class EmpleadoController extends Controller
      */
     public function store(EmpleadoRequest $request)
     {
-        if($request->has('nombre')){
-            $user=new User();
-            $user->nombre=$request->get('nombre');
-            $user->apellido=$request->get('apellido');
-            $user->domicilio=$request->get('domicilio');
-            $user->dni=$request->get('DNI');
-            $user->email=$request->get('email');
-            $user->tel=$request->get('tel');
-            $user->fecha_nac=$request->get('fecha_nacimiento');
-            $user->datos=$request->get('datos');
-            $user->tipo="empleado";
-            if($request->hasFile('foto')){
-                $file = $request->file('foto');
-                $path = 'img/fotos/';
-                $filename= time().'-'.$file->getClientOriginalName();
-                $uploadSuccess = $request->file('foto')->move($path , $filename);
-                $user->foto = $path . $filename;
+        if(auth()->user()->tipo === 'empleado'){
+            if($request->has('nombre')){
+                $user=new User();
+                $user->nombre=$request->get('nombre');
+                $user->apellido=$request->get('apellido');
+                $user->domicilio=$request->get('domicilio');
+                $user->dni=$request->get('DNI');
+                $user->email=$request->get('email');
+                $user->tel=$request->get('tel');
+                $user->fecha_nac=$request->get('fecha_nacimiento');
+                $user->datos=$request->get('datos');
+                $user->tipo="empleado";
+                if($request->hasFile('foto')){
+                    $file = $request->file('foto');
+                    $path = 'img/fotos/';
+                    $filename= time().'-'.$file->getClientOriginalName();
+                    $uploadSuccess = $request->file('foto')->move($path , $filename);
+                    $user->foto = $path . $filename;
+                }
+                $pass = Str::random(8);
+                $user->password=bcrypt($pass);
+                $user->save();
+                $emp=new Empleado();
+                $emp->user_id=$user->id;
+                $emp->puesto=$request->puesto;
+
+                $emp->idiomas=$request->idiomas;
+                $emp->user()->associate($user);
+                $emp->save();
+
+                $details=[
+                    'password' => $pass
+                ];
+
+                Mail::to($user->email)->send(new PassMail($details));
+
+                return back()->with('info','Se ha creado el registro.');
             }
-            $pass = Str::random(8);
-            $user->password=bcrypt($pass);
-            $user->save();
-            $emp=new Empleado();
-            $emp->user_id=$user->id;
-            $emp->puesto=$request->puesto;
-
-            $emp->idiomas=$request->idiomas;
-            $emp->user()->associate($user);
-            $emp->save();
-
-            $details=[
-                'password' => $pass
-            ];
-
-            Mail::to($user->email)->send(new PassMail($details));
-
-            return back()->with('info','Se ha creado el registro.');
-
-
-            //enviar email con contraseña care24web@gmail.com, webcare24
-
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
         }
     }
 
@@ -104,9 +114,14 @@ class EmpleadoController extends Controller
      */
     public function show($id)
     {
-        $empleado= Empleado::findOrFail($id);
+        if(auth()->user()->tipo === 'empleado'){
+            $empleado= Empleado::findOrFail($id);
 
-        return view('empleados.show', compact('empleado'));
+            return view('empleados.show', compact('empleado'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -116,9 +131,13 @@ class EmpleadoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
-        $empleado= Empleado::findOrFail($id);
-        return view('empleados.edit', compact('empleado'));
+    {    if(auth()->user()->tipo === 'empleado'){
+            $empleado= Empleado::findOrFail($id);
+            return view('empleados.edit', compact('empleado'));
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
+        }
     }
 
     /**
@@ -130,32 +149,36 @@ class EmpleadoController extends Controller
      */
     public function update(EmpleadoRequest $request, $id)
     {
-        $empleado= Empleado::findOrFail($id);
-        $empleado->user->nombre=$request->get('nombre');
-        $empleado->user->apellido=$request->get('apellido');
-        $empleado->user->domicilio=$request->get('domicilio');
-        $empleado->user->dni=$request->get('DNI');
-        $empleado->user->email=$request->get('email');
-        $empleado->user->tel=$request->get('tel');
-        $empleado->user->fecha_nac=$request->get('fecha_nacimiento');
-        $empleado->user->datos=$request->get('datos');
-        $empleado->user->fecha_baja=$request->get('fecha_baja');
-        if($request->hasFile('foto')){
-            $file = $request->file('foto');
-            $path = 'img/fotos/';
-            $filename= time().'_'.$file->getClientOriginalName();
-            $uploadSuccess = $request->file('foto')->move($path , $filename);
-            $empleado->user->foto = $path . $filename;
+        if(auth()->user()->tipo === 'empleado'){
+            $empleado= Empleado::findOrFail($id);
+            $empleado->user->nombre=$request->get('nombre');
+            $empleado->user->apellido=$request->get('apellido');
+            $empleado->user->domicilio=$request->get('domicilio');
+            $empleado->user->dni=$request->get('DNI');
+            $empleado->user->email=$request->get('email');
+            $empleado->user->tel=$request->get('tel');
+            $empleado->user->fecha_nac=$request->get('fecha_nacimiento');
+            $empleado->user->datos=$request->get('datos');
+            $empleado->user->fecha_baja=$request->get('fecha_baja');
+            if($request->hasFile('foto')){
+                $file = $request->file('foto');
+                $path = 'img/fotos/';
+                $filename= time().'_'.$file->getClientOriginalName();
+                $uploadSuccess = $request->file('foto')->move($path , $filename);
+                $empleado->user->foto = $path . $filename;
+            }
+            $empleado->user->save();
+            $empleado->puesto=$request->get('puesto');
+            $empleado->idiomas=$request->get('idiomas');
+
+            $empleado->save();
+            Session::flash('info', "Se han actualizado los datos.");
+
+            return redirect()->route('empleados.show', $id);
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
         }
-        $empleado->user->save();
-        $empleado->puesto=$request->get('puesto');
-        $empleado->idiomas=$request->get('idiomas');
-
-        $empleado->save();
-        Session::flash('info', "Se han actualizado los datos.");
-
-        return redirect()->route('empleados.show', $id);
-
     }
 
     /**
@@ -166,22 +189,27 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        $empleado= Empleado::findOrFail($id);
-        try
-        {
-            if(Servicio::where('empleado_id', '=', $id)->first() != null || Informe::where('empleado_id', '=', $id)->first() != null
-                || Incidencia::where('empleado_id', '=', $id)->first() != null){
-                return back()->with('error','Se han creado registros con este empleado, no puede ser eliminado.');
-            }else{
-                $empleado->user->delete();
-                $empleado->delete();
-                Session::flash('info', "Se ha eliminado el registro.");
-                return redirect()->route('empleados.index');
+        if(auth()->user()->tipo === 'empleado'){
+            $empleado= Empleado::findOrFail($id);
+            try
+            {
+                if(Servicio::where('empleado_id', '=', $id)->first() != null || Informe::where('empleado_id', '=', $id)->first() != null
+                    || Incidencia::where('empleado_id', '=', $id)->first() != null){
+                    return back()->with('error','Se han creado registros con este empleado, no puede ser eliminado.');
+                }else{
+                    $empleado->user->delete();
+                    $empleado->delete();
+                    Session::flash('info', "Se ha eliminado el registro.");
+                    return redirect()->route('empleados.index');
 
+                }
+
+            }catch(Exception $e){
+                return $e->getMessage();
             }
-
-        }catch(Exception $e){
-            return $e->getMessage();
+        }else{
+            Session::flash('danger','No está autorizado a acceder a esta ruta.');
+            return redirect()->route('inicio');
         }
     }
 }
